ppcls use new dir structure
